修复登陆不停开启MainActivity的bug

+++ app/src/main/java/com/ssthouse/moduo/view/activity/MainActivity.java
@@ -9,6 +9,7 @@ import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.AdapterView;
 import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.TextView;
@@ -49,6 +50,7 @@ public class MainActivity extends AppCompatActivity {
 
     private static final String EXTRA_IS_LOGIN_SUCCESS = "isLoginSuccess";
     private long exitTimeInMils = 0;
+    private boolean isLogOut = false;
 
     /**
      * 视频对话SDK管理类
@@ -115,10 +117,40 @@ public class MainActivity extends AppCompatActivity {
             tvOffline.setVisibility(View.VISIBLE);
         }
 
-        //TODO---应该是请求获取绑定了的设备才对---初始化设备listview
-//        deviceList = PreferenceHelper.getInstance(this).getDeviceList();
+        //请求获取绑定了的设备--初始化设备listview
         lvAdapter = new MainLvAdapter(this, deviceList);
         lv.setAdapter(lvAdapter);
+
+        lv.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                //TODO---尝试登陆
+                if (!deviceList.get(position).getXpgWifiDevice().isOnline()) {
+                    deviceList.get(position).getXpgWifiDevice().login(
+                            SettingManager.getInstance(MainActivity.this).getUid(),
+                            SettingManager.getInstance(MainActivity.this).getToken()
+                    );
+                }
+                Timber.e("clicked");
+            }
+        });
+
+        //TODO
+        findViewById(R.id.id_btn_fab).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                //TODO---尝试登陆
+                for (Device device : deviceList) {
+                    if (!device.getXpgWifiDevice().isOnline()) {
+                        device.getXpgWifiDevice().login(
+                                SettingManager.getInstance(MainActivity.this).getUid(),
+                                SettingManager.getInstance(MainActivity.this).getToken()
+                        );
+                    }
+                }
+                Timber.e("clicked");
+            }
+        });
 
         //初始化dialog
         waitDialog = new MaterialDialog.Builder(this)
@@ -298,7 +330,7 @@ public class MainActivity extends AppCompatActivity {
                 ToastHelper.show(this, "设备数据获取失败");
             }
         } else {
-            ToastHelper.show(this, "我不在最前..");
+            Timber.e("MainActivity不在最前..");
         }
     }
 
@@ -312,10 +344,6 @@ public class MainActivity extends AppCompatActivity {
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
-            case R.id.id_action_add_device:
-                //TODO---弹出添加设备dialog
-                showAddDeviceDialog();
-                break;
             case R.id.id_action_scan_device:
                 //TODO---扫码添加设备---在onActivityResult中处理回调
                 ScanUtil.startScan(this);
@@ -328,6 +356,15 @@ public class MainActivity extends AppCompatActivity {
                 XPGController.getInstance(this).getmCenter().cGetBoundDevices(
                         SettingManager.getInstance(this).getUid(),
                         SettingManager.getInstance(this).getToken());
+                break;
+            case R.id.id_action_log_out:
+                //TODO---登出
+                SettingManager.getInstance(this).clean();
+                //重新进入loading activity
+                LoadingActivity.start(this);
+                //结束当前activity
+                isLogOut = true;
+                finish();
                 break;
         }
         return super.onOptionsItemSelected(item);
@@ -396,7 +433,10 @@ public class MainActivity extends AppCompatActivity {
         super.onDestroy();
         EventBus.getDefault().unregister(this);
         communication.destory();//销毁sdk
-        android.os.Process.killProcess(android.os.Process.myPid());//确保完全退出，释放所有资源
+        //todo---如果不是跳转到登陆界面才杀死进程
+        if (!isLogOut) {
+            android.os.Process.killProcess(android.os.Process.myPid());//确保完全退出，释放所有资源
+        }
     }
 
     @Override
